requirements.txt
marvin==2.3.5
fastapi
requests==2.31.0
sentry_sdk==2.7.1
python-dotenv==1.0.0
openai==1.35.7
firecrawl-py
celery==5.4.0
posthog
watchdog
pydantic
redis==5.0.7
googlemaps==4.10.0
langchain-community==0.2.6
docarray==0.40.0
langchain-openai==0.1.14
jq==1.8.0a2
playwright==1.41.0
duckduckgo-search==6.1.9
jupyter==1.0.0
supabase==2.5.1